requirements.txt
aiohttp==3.7.3
aiomysql==0.0.21
aioredis==1.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
chardet==3.0.4
click==7.1.2
ecdsa==0.14.1
greenlet==1.1.2
h11==0.12.0
hiredis==2.0.0
idna==3.3
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
pyasn1==0.4.8
pycryptodome==3.10.1
pydantic==1.8.2
PyMySQL==0.9.3
python-jose==3.2.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.23
sqlalchemy-stubs==0.4
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.12.3
yarl==1.7.2
fastapi==0.73.0
